Move azure service fabric tests to it's own project
